refactor: rename model files from kebab-case to camelCase
api-resource.js → apiResource.js
event-resource.js → eventResource.js
entity-type.js → entityType.js
