+++ lib/model/entityType.js
