+++ lib/model/eventResource.js
@@ -2,7 +2,7 @@ const { ensureAccessStrategies } = require("../access-strategies");
 const { ORD_RESOURCE_TYPE, RESOURCE_VISIBILITY } = require("../constants");
 const { determineVisibility } = require("./visibility");
 const { stripNamespacePrefix, findPackageId } = require("./naming");
-const { mergeResourceProperties } = require("./api-resource");
+const { mergeResourceProperties } = require("./apiResource");
 
 function createEventResources(service, config, packageIds, accessStrategies) {
     const { extensions, definition } = service;
