+++ lib/model/index.js
@@ -1,6 +1,6 @@
-const { createApiResources, mergeResourceProperties, buildResourceDefinitions } = require("./api-resource");
-const { createEventResources } = require("./event-resource");
-const { createEntityTypes, hasSAPPolicyLevel } = require("./entity-type");
+const { createApiResources, mergeResourceProperties, buildResourceDefinitions } = require("./apiResource");
+const { createEventResources } = require("./eventResource");
+const { createEntityTypes, hasSAPPolicyLevel } = require("./entityType");
 const { createGroup } = require("./group");
 const { createProducts, createPackages, createConsumptionBundles } = require("./package");
 const { determineVisibility, isPrimaryDataProductService } = require("./visibility");
