+++ lib/model/package.js
@@ -1,7 +1,7 @@
 const _ = require("lodash");
 const defaults = require("../defaults");
 const Logger = require("../logger");
-const { hasSAPPolicyLevel } = require("./entity-type");
+const { hasSAPPolicyLevel } = require("./entityType");
 
 function createProducts(config) {
     const products = defaults.products(config.packageName);
